requirements.txt
django~=5.0
djangorestframework~=3.14.0
celery[redis]~=5.3.0
redis~=5.0.1
requests~=2.31.0
ujson~=5.9.0
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
psycopg2-binary>=2.9.0
setuptools>=65.0.0
drf-yasg~=1.21.0python-decouple~=3.8
drf-yasg~=1.21.0
black~=23.12.0
flake8~=7.0.0
isort~=5.13.0
